module system: rewrite
