+++ bin/pm2
@@ -516,10 +516,11 @@ commander.command('uninstall <module>')
 
 
 commander.command('publish')
+  .option('--tarball', 'is local tarball')
   .alias('module:publish')
   .description('Publish the module you are currently on')
-  .action(function() {
-    pm2.publish();
+  .action(function(opts) {
+    pm2.publish(opts);
   });
 
 commander.command('set [key] [value]')
