+++ lib/API/Modules/TAR.js
@@ -10,42 +10,24 @@ var path = require('path');
 var fs = require('fs');
 var os = require('os');
 var spawn = require('child_process').spawn;
+var exec = require('child_process').exec;
 
 module.exports = {
   install,
   uninstall,
-  start
+  start,
+  publish
 }
 
 /**
  * Module management to manage tarball packages
  *
- * pm2 install http.tar.gz --tarball
+ * pm2 install http.tar.gz
  * pm2 uninstall http
  *
  * - the first and only folder in the tarball must be called module (tar zcvf http module/)
  * - a package.json must be present with attribute "name", "version" and "pm2" to declare apps to run
  */
-
-/**
- * Uncompress only module/package.json and retrieve the "name" attribute in the package.json
- */
-function getModuleName(module_filepath, cb) {
-  var install_instance = spawn('tar', ['zxf', module_filepath, '-C', os.tmpdir(), `${cst.MODULE_BASEFOLDER}/package.json`], {
-    stdio : 'inherit',
-    env: process.env,
-		shell : true
-  })
-
-  install_instance.on('close', function(code) {
-    try {
-      var pkg = require(path.join(os.tmpdir(), `${cst.MODULE_BASEFOLDER}/package.json`))
-      return cb(null, pkg.name)
-    } catch(e) {
-      return cb(e)
-    }
-  });
-}
 
 function install(CLI, module_filepath, opts, cb) {
   Common.log(`Unpacking local tarball ${module_filepath}`)
@@ -166,5 +148,78 @@ function uninstall(PM2, module_name, cb) {
     PM2._operate('deleteProcessId', `${module_name}:${app.name}`, () => next())
   }, () => {
     cb(null)
+  })
+}
+
+
+/**
+ * Uncompress only module/package.json and retrieve the "name" attribute in the package.json
+ */
+function getModuleName(module_filepath, cb) {
+  var install_instance = spawn('tar', ['zxf', module_filepath, '-C', os.tmpdir(), `${cst.MODULE_BASEFOLDER}/package.json`], {
+    stdio : 'inherit',
+    env: process.env,
+		shell : true
+  })
+
+  install_instance.on('close', function(code) {
+    try {
+      var pkg = require(path.join(os.tmpdir(), `${cst.MODULE_BASEFOLDER}/package.json`))
+      return cb(null, pkg.name)
+    } catch(e) {
+      return cb(e)
+    }
+  });
+}
+
+function publish(opts, cb) {
+  var pkg = require(path.join(process.cwd(), 'package.json'))
+
+  if (!pkg.name) throw new Error('Attribute name should be present')
+  if (!pkg.version) throw new Error('Attribute version should be present')
+  if (!pkg.pm2 && !pkg.apps) throw new Error('Attribute apps should be present')
+
+  var current_path = process.cwd()
+  var base_folder = path.dirname(current_path)
+  var module_name = path.basename(current_path)
+  var archive_path = path.join(os.tmpdir(), 'module.tar.gz')
+
+  process.chdir(os.tmpdir())
+
+  Common.log('Creating package')
+  var cmd = `tar zcf module.tar.gz -C ${base_folder} --transform 's,${module_name},module,' ${module_name}`
+
+  var tar = exec(cmd, (err, sto, ste) => {
+    if (err) {
+      console.log(sto.toString().trim())
+      console.log(ste.toString().trim())
+    }
+  })
+
+  tar.on('close', function () {
+    Common.log('Package created')
+
+    var bitmap = fs.readFileSync(archive_path)
+    var module = Buffer.from(bitmap).toString('base64')
+
+    pkg.id = pkg.name
+    pkg.module = module
+
+    var uri = 'http://localhost:9003/api/v1/modules'
+    Common.log(`Sending Package to remote ${uri}`)
+
+    require('needle')('post', uri, pkg)
+      .then(resp =>  {
+        if (resp.statusCode !== 201) {
+          Common.err(`${pkg.name}-${pkg.version}: ${resp.body.msg}`)
+          process.exit(1)
+        }
+        Common.log(`Module ${module_name} published under version ${pkg.version}`)
+        process.exit(0)
+      })
+      .catch((e) => {
+        Common.err(e)
+        process.exit(1)
+      })
   })
 }
