+++ lib/API/Modules/index.js
@@ -54,10 +54,10 @@ module.exports = function(CLI) {
   /**
    * Publish module on NPM + Git push
    */
-  CLI.prototype.publish = function(module_name, cb) {
+  CLI.prototype.publish = function(opts, cb) {
     var that = this;
 
-    Modularizer.publish(function(err, data) {
+    Modularizer.publish(opts, function(err, data) {
       if (err)
         return cb ? cb(Common.retErr(err)) : that.speedList(cst.ERROR_EXIT);
       return cb ? cb(null, data) : that.speedList(cst.SUCCESS_EXIT);
